First screen API Integration complete | Favourite cities feature also added

+++ app/src/main/java/com/example/muhbandtekamshuru/demoapp/data/network/model/Clouds.java
@@ -8,9 +8,9 @@ public class Clouds {
 
     @SerializedName("all")
     @Expose
-    private int all;
+    private double all;
 
-    public int getAll() {
+    public double getAll() {
         return all;
     }
